+++ driver-core/CHANGELOG.rst
@@ -3,8 +3,10 @@ CHANGELOG
 
 2.0.8:
 ------
+
 - [bug] Fix token awareness for case-sensitive keyspaces and tables (JAVA-526)
 - [bug] Check maximum number of values passed to SimpleStatement (JAVA-515)
+- [improvement] Expose the driver version through the API (JAVA-532)
 
 
 2.0.7: